# whatsapp-ui/src/app/services/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserResponse {
  email?: string;
  firstName?: string;
  id?: string;
  lastName?: string;
  lastSeen?: string;
  online?: boolean;
}